Introduce DefaultMavenDeployRetrier unit tests
