+++ subprojects/maven/src/main/java/org/gradle/api/publication/maven/internal/action/MavenDeployRetrier.java
@@ -16,73 +16,8 @@
 
 package org.gradle.api.publication.maven.internal.action;
 
-import org.apache.http.conn.HttpHostConnectException;
-import org.apache.maven.artifact.ant.RemoteRepository;
-import org.gradle.internal.UncheckedException;
-import org.gradle.internal.resource.transport.http.HttpErrorStatusCodeException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.sonatype.aether.deployment.DeploymentException;
 
-import java.net.SocketTimeoutException;
-import java.util.concurrent.Callable;
-
-public class MavenDeployRetrier {
-    private static final Logger LOGGER = LoggerFactory.getLogger(MavenDeployRetrier.class);
-    private final RemoteRepository remoteRepository;
-    private final Callable operation;
-    private final int maxDeployAttempts;
-    private final int initialBackOff;
-
-    public MavenDeployRetrier(Callable operation, int maxDeployAttempts, int initialBackOff, RemoteRepository remoteRepository) {
-        this.operation = operation;
-        this.maxDeployAttempts = maxDeployAttempts;
-        this.initialBackOff = initialBackOff;
-        this.remoteRepository = remoteRepository;
-    }
-
-    public void deployWithRetry() throws DeploymentException {
-        int backoff = initialBackOff;
-        int retries = 0;
-        while (retries < maxDeployAttempts) {
-            retries++;
-            Throwable failure;
-            try {
-                operation.call();
-                if (retries > 1) {
-                    LOGGER.info("Successfully deployed resource after {} retries", retries - 1);
-                }
-                break;
-            } catch (Exception throwable) {
-                failure = throwable;
-            }
-            boolean doNotRetry = !isLikelyTransientNetworkingIssue(failure);
-            if (doNotRetry || retries == maxDeployAttempts) {
-                throw new DeploymentException("Could not deploy to remote repository | " + failure.getMessage(), failure);
-            } else {
-                LOGGER.info("Error while accessing remote repository {}. Waiting {}ms before next retry. {} retries left", remoteRepository, backoff, maxDeployAttempts - retries, failure);
-                try {
-                    Thread.sleep(backoff);
-                    backoff *= 2;
-                } catch (InterruptedException e) {
-                    throw UncheckedException.throwAsUncheckedException(e);
-                }
-            }
-        }
-    }
-
-    private static boolean isLikelyTransientNetworkingIssue(Throwable failure) {
-        if (failure instanceof SocketTimeoutException || failure instanceof HttpHostConnectException) {
-            return true;
-        }
-        if (failure instanceof HttpErrorStatusCodeException) {
-            HttpErrorStatusCodeException httpError = (HttpErrorStatusCodeException) failure;
-            return httpError.isServerError();
-        }
-        Throwable cause = failure.getCause();
-        if (cause != null && cause != failure) {
-            return isLikelyTransientNetworkingIssue(cause);
-        }
-        return false;
-    }
+interface MavenDeployRetrier {
+    void deployWithRetry() throws DeploymentException;
 }
